refactor syntax
