+++ src/semantic/BaseVisitor.ts
@@ -1,7 +1,7 @@
 import { CstNode, IToken } from "chevrotain";
 import { baseVocabulary } from "../lexer";
 import { BaseParser } from "../parser";
-import { AbstractTree, CronExpression, Expression, intervalExpr, rangeExpr, StringLiteral } from "../syntax/BaseSyntax";
+import { AbstractTree, CronExpression, Expression, intervalExpr, rangeExpr, StringLiteral } from "../syntax/base";
 
 const baseVisitorConstructor = new BaseParser(baseVocabulary).getBaseCstVisitorConstructor();
 export class BaseVisitor extends baseVisitorConstructor {
