+++ src/syntax/base/CronExpression.ts
@@ -1,3 +1,4 @@
+import { everyExpr } from ".";
 import AbstractTree from "./AbstractTree";
 
 export default class CronExpression extends AbstractTree {
